Avoid zombie processes.
Since the warmup subprocess waits for Zope to be finished starting
up and handling requests, we should not block the startup.
But it is important to collect the result of the subprocess (by
using ``subprocess.check_call``) in order to avoid zombie processes.
However, since ``subprocess.check_call`` is blocking the caller, we call
it within a new thread which will have the patience to wait for the
subprocess to finish and the main startup process can continue.